requirements.txt
tensorflow==2.13.0
opencv-python==4.8.0.74
streamlit==1.26.0
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
numpy==1.24.3
pillow==9.5.0
scikit-learn==1.3.0
pympler==1.0.1
tzlocal==4.3
